Handle upstream's versioning
Upstream embeds version info in the generated binaries via the VERSION file,
which is generated by dubious means. Instead, we'll embed it in debian/rules
and handle it from there. This isn't ideal, as it requires that we keep rules
in sync with changelog by hand, but the alternatives are worse.